+++ src/android/geolocation-activity.d.ts
@@ -1,0 +1,4 @@
+export interface IGeolocationActivity {
+    geolocationCallback: android.webkit.GeolocationPermissions.ICallback;
+    geolocationOrigin: string;
+}
