+++ src/android/geolocation-activity.ts
@@ -1,0 +1,4 @@
+export interface IGeolocationActivity {
+    geolocationCallback: android.webkit.GeolocationPermissions.ICallback;
+    geolocationOrigin: string;
+}
