+++ src/android/index.ts
@@ -1,0 +1,2 @@
+export { IGeolocationActivity } from './geolocation-activity';
+export { IUploadFileActivity } from './upload-file-activity';
