+++ src/android/nativechat-web-chrome-client.d.ts
@@ -1,0 +1,5 @@
+export declare class NativeChatWebChromeClient extends android.webkit.WebChromeClient {
+    onGeolocationPermissionsShowPrompt(origin: string, callback: android.webkit.GeolocationPermissions.ICallback): void;
+    onShowFileChooser(webview: android.webkit.WebView, filePathCallback: android.webkit.ValueCallback<android.net.Uri>, fileChooserParams: any): boolean;
+    shouldOverrideUrlLoading(webview: android.webkit.WebView, url: string): boolean;
+}
