+++ src/android/upload-file-activity.ts
@@ -1,0 +1,3 @@
+export interface IUploadFileActivity {
+    uploadCallback: android.webkit.ValueCallback<android.net.Uri>;
+}
