+++ src/index.ts
@@ -1,1 +1,4 @@
-export { NativeChat, NativeChatConfig, Session, User, IGeolocationActivity, IUploadFileActivity } from './nativechat';
+export { NativeChat } from './nativechat';
+export { NativeChatConfig, Session, User } from './models';
+export { IGeolocationActivity, IUploadFileActivity } from './android';
+
