+++ src/models/nativechat-config.ts
@@ -1,0 +1,11 @@
+import { Session } from "./session";
+import { User } from "./user";
+
+export interface NativeChatConfig {
+    botId: string;
+    channelId: string;
+    channelToken: string;
+    gtmId?: string;
+    session?: Session;
+    user?: User;
+}
