+++ src/models/session.ts
@@ -1,0 +1,5 @@
+export interface Session {
+    clear?: boolean;
+    context?: object;
+    userMessage?: string;
+}
