+++ src/models/user.d.ts
@@ -1,0 +1,4 @@
+export interface User {
+    id?: string;
+    name?: string;
+}
