+++ src/models/user.ts
@@ -1,0 +1,4 @@
+export interface User {
+    id?: string;
+    name?: string;
+}
